[tests] Move --device argument up a level
When trying to run the unit tests from the top level directory, was getting:
```
ValueError: option names {'--device'} already added
```
since this was defined in both `operators` and `language` tests.
